1.0.4 added services, repositories, exceptions

+++ src/main/java/com/insurancemanager/repository/PolicyRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PolicyRepository extends JpaRepository<Policy, Integer> {
+public interface PolicyRepository extends JpaRepository<Policy, Long> {
     Page<Policy> findAll(Pageable pageable);
 }
